requirements.txt
appdirs==1.4.4
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
multitasking==0.0.11
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pandas-datareader==0.10.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.12
yfinance==0.1.77
